+++ changelog.md
@@ -4,6 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 # [Unreleased](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v7.1.0...HEAD)
+### Added
+* [PCM-1279](https://inindca.atlassian.net/browse/PCM-1279) – calling `sdk.setAccessToken()` will now also pass
+    that token down to the streaming-client and the client-logger.
+
 
 # [v7.1.0](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v7.0.0...v7.1.0)
 ### Fixed