+++ changelog.md
@@ -3,7 +3,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [Unreleased](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v7.1.0...HEAD)
+# [Unreleased](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v7.1.1...HEAD)
+# [v7.1.1](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v7.1.0...v7.1.1)
+### Changed
+* [PCM-1387](https://inindca.atlassian.net/browse/PCM-1387) - Made screenRecordingMetadata a required param for acceptSession for screen recording sessions; Those metadatas get sent to the server automatically.
 
 # [v7.1.0](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v7.0.0...v7.1.0)
 ### Fixed